% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MainAnalysis.R
\name{execute}
\alias{execute}
\title{Execute OHDSI Study 2}
\usage{
execute(dbms, user, password, server, port = NULL, cdmSchema, resultsSchema,
  minCellCount = 1, sourceName = "source_name", folder = getwd())
}
\arguments{
\item{dbms}{The type of DBMS running on the server. Valid values are
\itemize{
  \item{"mysql" for MySQL}
  \item{"oracle" for Oracle}
  \item{"postgresql" for PostgreSQL}
  \item{"redshift" for Amazon Redshift}
  \item{"sql server" for Microsoft SQL Server}
}}

\item{user}{The user name used to access the server.}

\item{password}{The password for that user}

\item{server}{The name of the server}

\item{port}{(optional) The port on the server to connect to}

\item{cdmSchema}{Schema name where your patient-level data in OMOP CDM format resides}

\item{resultsSchema}{Schema where you'd like the results tables to be created (requires user to have create/write access)}

\item{minCellCount}{The smallest allowable cell count, 1 means all counts are allowed}

\item{sourceName}{Short name that will be appeneded to results table name}

\item{folder}{The name of the local folder to place results;  make sure to use forward slashes (/)}
}
\value{
Study results are placed in CSV format files in specified local folder.
}
\description{
Execute OHDSI Study 2
}
\details{
This function executes OHDSI Study 2 -- Treatment Pathways Study
Protocol 12 months.  This is a study of treatment pathways in hypertension,
diabetes and depression during the first 12 months after diagnosis.
Detailed information and protocol are available on the OHDSI Wiki.
}

